Added Table updates for #5
Removed the TableHeaderRow component for using a TableHeader component follwed by a TableRow component

Renamed the TableData component to TableColumn. It can now be a td or a th based on header={BOOL}.

Updated the TableRow component to automatically justify smallest columns to include at least 56px between columns and autogrow the biggest column.