Implemented #25
The select field now allows for a new item to be selected by pressing a letter or number
to select an item beginning with that letter or number.

The user can press the up or down arrow key to select the next or previous item.

Pressing tab will advance to the next tabbable item on the page and not the menu items.