Snackbar Fixes
Updated the snackbar to allow for better chained toasts. The timeout for the enter
needs to be doubled when it is chained because there is a 450ms delay on chained toasts.

Cleared timeouts on chained toasts if a user manually dismissed it before the auto timeout.

Fixed it for Toast test.

Removed the padding reduction on the flat buttons in toasts.