Fixed Props for TextField
Fixed the props for TextField that only style and className are applied to the surrouding div. The remaining
props are now applied to the input instead. This is to help with some of the missing props of name and id and all the
many events that can be attached to a text field.

Fixed a css bug where active was being colored when it should have been on focus.