Small Styling Fixes for NavigationDrawer
Added a new prop for the Navigation Drawer
so that the header in the Drawer can be *fixed*
to the top. This sets the md-list underneath the drawer
to be scrollable and sets it's max height relative to the
current device height.

Also fixed some of the places where I used `%` instead of `vw` or `vh`.