Fixed the Double Icons and Disabled Button
The DatePickerCalendar component was displaying two FontIcons after the
changes for how buttons work. Updated it so it follows the same idea of
children and iconClassName for nested icons.

Fixed the stypes for buttons so that a disabled button will no longer
trigger the click event when clicking on a div or i tag.