Updated DatePicker for some Keyboard Accessibility
The Datepicker can now be opened with by focusing the input and pressing
the enter key. When the calendar is opened, the current date will now be
foucsed instead of the toggle for the year picker. #173

When the year picker is opened, the active year will focused by default.

This isn't really good keyboard accessibility though. The _inline_ date
picker should really be three select dropdowns that allows a user to
select, day, month, year.