Updated TimePicker for some Keyboard Accessibility
Updated the TimePicker so that it could be opened by tab focusing the
input field, and then pressing enter. #173

Also updated the the ClockTime components to be able to be tab selected.
This still isn't really ideal for keyboard users. This should really be
a dropdown or something for them.