Updated Portal to render inline when in portals
When a Dialog is rendered in a PureComponent, the automatic setting of
the renderNode does not get persisted since it relies on context. There
is now a fallback that if any Portal is rendered within another portal,
it will be rendered inline instead of as the first child in the body.

This functionality should all be removed in v1.1.0 since most comonents
stopped using the portal by default.

Closes #229