Updated Layover logic for viewport filling
The Layover can now attempt to fill the entire viewport (especially
helpful for mobile devices). This fixes most of the remaining problems
for the mobile devices when the Autocomplete is in the main toolbar.

The _next_ step is to also keep items in viewport based on their
position on the page. sigh.