Update Inks for Snapshot Testing
When snapshotting a component with the react-test-renderer and a
TransitionGroup, it appears to fail when a child is not a full React
Component Class that has the lifecycle methods.

Instead of having a nice span to get dom elements, I had to fall back to
using the findDOMNode API to add event listeners.

Closes #611