Updated DialogContent sizing logic
There were some cases where the height and scrollHeight were always the
same in the dialog content so that the max-height setting was never
triggered even though there was more content than the entire viewport.

Updated the logic so the calculations always happen, but setting the
state is still dependent on if the max-height differs.