Fixed TextField Floating Labels when Controlled
The TextField floating label will correctly still be floating when the
user removes all text from a text field while the value is controlled by
props. Before this patch, the label would stop floating once the value
prop is blank even if the user is still focused on the text field.