Updated Collapse and removed withMountingTransition
I ended up removing the `withMountingTransition` since the new API for
react-transition-group is a lot nicer than in the past, and I will be
leveraging it when more transitions need to be made.

Updated the Collapse component to work a lot better when the user
quickly expands and collapses the child element. Also wrote some tests
for the Collapse. This _should_ have been made with the
react-transition-group, but I couldn't get the dynamic style application
to occur at a good time that allowed for fluid transitions. :/ I'll
probably have to come back to this.