Added IconRotator and small icon changes
Added an IconRotator component to easily rotate icons (normally for
expansion areas).

Updated svg icons so that the child elements have no pointer events so
that it's easier to determine if an icon was clicked since it'll just be
the svg with a "rmd-icon" className instead of any element.

Updated TextIconSpacing so that when it is wrapped, the container span
will center align with flex to make things look nicer.