Added a color fix for buttons and temp interaction states
Added temp interaction states to the button component so the user gets
feedback when they hover, focus, or press the button.

With these changes, I also added a workaround so that the button content
doesn't get affected by the opacity applied to these interaction states.

Updated the default background color to be the old background color.