Updated the docs to build a bit quicker
I'll need to look into this a bit more, but it looks like with the
current import architecture for the `@react-md` stuff (maybe it's
because of the inclusing of node_modules?) the initial build is
extremely slow since it imports every file from react-md in.

I now pre-compile the base styles to a app.css file and just import that
instead of letting webpack compile and include the scss file. Startup
time changed from about 40seconds-2min to 10 seconds...