Fixed the Portal component
I didn't really understand this at the time, but it doesn't look like
you should be conditionally using the `createPortal` API as this causes
problems with other libraries. Not sure whose _fault_ it is, but if you
switch between using `createPortal` and `null`, it'll break the
CSSTransition component and prevent animations since it will start
applying `undefined` as a classname or just unmounting immediately.

The current fix is to always render the portal in the DOM and rely on
the visibility toggle to be handled by the children instead.