Portal, Overlay, Elevation, and Scroll Lock Changes
Fixed the overlay so that it doesn't lose the `rmd-overlay--active`
className if it gets re-rendered after finishing the animation.

Fixed the Portal component to not throw an error immediately after your
app has been SSRed

Updated the elevation pseudo-element transition to include a z-index.

Updated the useScrollLock hook to quick escape if disabled instead of
checking the dom and then escaping.