Progress Package Updates
Updated the linear and circular progress components to be rendered in
span instead of div so that they can be rendered within buttons.

Added the determinate examples for progress.

Added examples for using the progress components within buttons.

Updated the circular progress to use scss variables for sizing and a
theme variable so that it is easier to style.

Updated the circular progress to be display inline-flex and flex instead
of inline-block and block. This was just so it can display nicely in
buttons/TextIconSpacing components