Added some dialog, wia-aria and state fixes
Updated the dialog to pass props down to the FocusContainer and updated
the prop interfaces to reflect this as well.

Updated the FocusContainer to export another prop interface for
extending this component's functionality.

Updated the keyboard click polyfill so that it isn't triggered on button
elements since it would click the button twice.

Updated the focusable elements cache to include the current instance if
there are no items and it is focusable itself instead of throwing an
error.