Finished Up Tooltip State Work Round 1
Finished making the logic for how to show and hide a tooltip based on
the user interaction type.

Wrote a few base tests to prevent type checking from failing for the
tooltip states, but need a lot more.

Deleted the unneeded files.

Added some comments and @private tags for some utils/hooks