More Documentation Changes
Created a custom server so favicon.ico and robots can be served from
root.

Added support for cookies to the custom server so the theme can be
rendered server side.

Updated the Google Fonts definitions since there is a new display=swap
functionality. Unfortunately can't use it yet since it breaks in
firefox?

Moved the analytics page out of the `pages` folder since it accidentally
made a route for it.