More Menu Changes
Removed the onItemClick prop since it is pretty much worthless

Updated the defaultItemRenderer to return React elements instead of
wrapping them in a MenuItem. This allows for custom menu items to be
passed in with their click handlers already defined.

Created a variable for the icon spacing within menus since you don't
need the same spacing as lists to match keylines.

Updated the menu documentation for these new updates.