Lots of changes that should've been multi-commits
Updated the app-bar package to export a `useActionClassName` hook and
interface so that the `DropdownMenu` can be rendered as an app bar action.

Updated the `DropdownMenu` to be renderable as the `AppBarAction`
component when the `first`, `last`, `inheritColor` or `asAppBarAction`
props are enabled.

Fixed the spelling of `menuLabelledBy`.

Updated the `DropdownMenu` to pass the portal props down into the
menuRenderer.

Udpated the `DropdownMenu` to remove the old `onItemClick` prop.

Updated the `DropdownMenu` and Menu components to pass down the
`positionOptions` so that the `useFixedPositioning` can be configurable.

Updated the `DropdownMenu` to default to setting the `aria-labelledby`
to the `MenuButton`'s id.

Updated the `Menu` component to correctly close when the tab key is
pressed.

Updated the `Menu` component to not trigger the `onRequestClose`
callback when its `MenuButton` is clicked since the `MenuButton` click
handler already toggles the visibility.

Added a `MenuItemLink` component and updated the `menuItemRenderer` to
support rendering links within menus.

Fixed the `ListItemLink` to no longer require an id (first
implementation) and correctly pass the `className` prop down to be
merged.

Updated the icon package to just have a `rmd-icon-text-spacing` mixin
instead of the before/after variants.

Updated the `useFixedPositioning` and `getFixedPosition` so that the
`getFixedPosition` always returns `position: fixed` and moved the
`transformOrigin` logic into it.

Added more "fixes" to the `getFixedPosition` util.

Updated the main watcher to also build cjs and auto-sandbox empty
sandboxes.

Updated the documentation site's main layout to use an overflow menu on
mobile.

Added a lot more menu documentation.