Updated FixedPositioning logic for VH Overflow
There are some cases where it can be nice to prevent the fixed element
to be contained within the viewport and instead rely on the full page
scroll to show the rest of the content. This works fairly well, but
scrolling quickly to the bottom will look a bit weird since the fixed
positioning logic is normally throttled on scroll for performance. It's
probably a bit worse on Mac since you can normally scroll past the
"real" bottom of the page.