New Keyboard Movement (breaks Menus)
This added a lot of new functionality and features and probably should
have been multiple commits... but oh well.

Added:

A new `scrollIntoView` util that will conditionally scroll an element
into view of another container element. This will work even if the
container scrollable element does not have `position: relative` set
unlike the `el.offsetTop`.

The `useActiveDescendantMovement` hook to handle keyboard movement when
the focused DOM element should not change, but other elements should be
"focused" within the DOM for more complex widgets.

The `useFocusMovement` hook to handle true DOM focus keyboard movement.

`MovementPreset` presets so that using the two hooks above can be even
easier by choosing an existing supported keyboard movement
configuration.

Fixed:

The top position in the createVerticalPosition helper.

Updated:

The loop util to be able to not "loop" and instead stop at the min and
max values provided for a general a11y helper in keyboard movement.

The `useKeyboardMovement` util to be more of a "helper" private hook to
be used by the `useActiveDescendantMovement` and `useFocusMovement`
hooks.

Removed:

The some duplicated files between wia-aria and the search folders after
the moves.