Created a ScaleTransition for re-use
Created a ScaleTransition that is basically the menu animation that does
scale and opacity changes. Also added a scale-y transition that only
does the scaleY version instead of scale.

Updated the @react-md/menu package to use the ScaleTransition component
which also fixed an issue where I forgot to install
react-transition-group in that package anyways...