Added Select, Listbox, and Option components
Updated the `FloatingLabel` component to be able top render as a span
for the custom select implementation.

Added a simple Option component that is a wrapper for the
`SimpleListItem` with additional accessibility for a listbox option

Added a Listbox component that can be rendered inline or temporarily
with a select button.

Added a Select component that is an accessible and more style-able
version of the <select> element.

Updated the useFocusState private hook to also support divs.