Select documentation and fixed
Updated the Select/Listbox documentation to show some more use-cases for
customizing the options and display labels within each option as well as
within the Select's button.

Added the ability to customizing the Select's button value and
optionally show the option's `letIcon` or `leftAvatar` within the
Select's button as well.