Updated fixed positioning for another width option
The fixed positioning can now set the minWidth instead of width option
if desired so that the fixed element is at least the same width as the
container/fixed to element.

This also changed the styles for the equalWidth so that it uses the
width property instead of right. I couldn't get the right value
calculated correctly and was off by 16px or so sometimes.