Updated fixed positioning onScroll and AutoComplete
The onScroll behavior for the useFixedPositioning will now return an
object of the element, fixedTo, and visible instead of multiple
arguments. The visible flag will determine if the fixedTo element is
still visible within the viewport.

Updated the AutoComplete to no longer hide on scroll by default, but
instead only hide when the input is no longer in view within the viewport