Initial Tree rewrite commit
Started working on redoing the tree package API and rewrite for the new
keyboard movement and fixing some of the API.

This added some initial tests and some of the new utils and types.

The styles are mostly copied over from @react-md/tree, but also added
support for the focused item styles.