Fixed Possible Security Bug in Doc Site
Updated the "Go Back" button on the CodePreview sandbox page to no
longer go back to `from` urls that do not start with `/packages`. If
someone _really_ wanted to send a malicious link with this website (no
idea why they'd choose this...), they could send a link with a
"malicious" url as the from url param and have it redirect there
afterwards.

So for example, the url for the Multi Select Tree example would be:
https://react-md.dev/sandbox?pkg=Tree&name=Multi%20Select%20Tree&from=%2Fpackages%2Ftree%2Fdemos%23multi-select-tree-title

A mean user could change the from value to their website so if the user
hits the "Go Back" button, it would redirect there instead:

https://react-md.dev/sandbox?pkg=Tree&name=Multi%20Select%20Tree&from=http%3A%2F%2Fi-am-a-bad-website.com

So if the from query param does not start with `/packages` (something
guarenteed to be on the same domain), just go back to `/sandbox`
instead.

Updated some documentation and types within the documentation site.

I also added testing to the documentation package and updated the travis
config to run the tests there as well.