Tree fixes and added visibleIndex functionality
Fixed the flattened tree stuff to use the provided `rootId` instead of
only using null.

Updated the types for the id+ref object to be metadata and also include
the visible index of the item. This allows for drag and drop
functionality from react-beautiful-dnd even though they don't support
nested lists at the time of writing this.

ALso updated the TreeItem component to have a `liRef` prop which is now
the ref to the `HTMLLIElement`. The forwarded ref will be applied to the
span/link component instead since it's normally what you want because
the `<li>` also contains the sub-groups (if open).