Revert "Revert "Upgraded nextjs for src folder support""
This reverts commit 6807fff031d1af58f5c7a45cb2082452b1010832.

I was stupid and forgot to update the `copy-readmes` script to copy into
the src directory as well. It was still copying into the root `pages`
folder so since the `<rootDir>/pages` folder exists (even though there
was still `<rootDir>/src/pages`), the first one would be used and would
result in an empty app.