Initial Sheet Rework
This implemented most of the new rework for the Sheet component so that
it uses the dialog role. This will need to be re-worked a little bit so
that it doesn't need to be rendered as a dialog (for inline stuff)  and
will be finished once the layout package gets started.