Fixed Select form submit
I didn't know that triggering a form.submit() is actually different than
a type="submit" action in that the onsubmit handler isn't called with
form.submit(). This means that the Select component prevented ajax forms
from working correctly if the user pressed enter while focusing it.

This workaround will try to find a submit button and trigger a click
event on that instead. If a submit button does not exist, temporarily
create one in the form and click it.