Dependency Fixes and Tree Renderer Update
Fixed some missing dependencies in the documentation package

Updated the react-md package to also export all icons from
material-icons.

Updated the defaultTreeItemRenderer to also automatically pass down the
`as` attribute on each tree item to work out of the box with nextjs.

Updated the routesTree component to render dynamic routes with nextjs.