Mobile Documentation Fixes
Decreased the ridiculous font size for headline-1, headline-2, and
headline-3 for mobile devices. No idea why these are the defaults. I
wonder if the documentation has changed for these?

Also fixed the Layout in the documentation site to mount the
TOCVisibilityProvider before the Layout component so that the action
menu on mobile (which is rendered in the Layout header) can gain the
context correctly.