Updated to use no-unused-vars eslint rule again
Added the @typescript-eslint/no-unused-vars rule back. This actually
caught a lot of "errors" for variables I didn't actually mean to not
use.

Also renamed the copy-readmes script to just be readmes

Removed the sandbox --empty to prebuild for docs since the sandboxes are
now committed and will never be missing.