Added Additional Demo and Phone Config
Added a new emulated option for the Demos so that the demo can be
"emulated" in a phone without needing to include the phone source in the
sandboxes.

Added a new contentStacked option so that the emulated phone will use
the display: flex and flex-direction: column styles so the app bar
doesn't need to be position: fixed with correct offset and scrolling.

Fixed the transition package to correctly depend on
react-transition-group