Fixed useResizeObserver for cleaner unmounts
Since it's sometimes helpful to use arrow functions to get an element,
it ended up causing the onResize callback to be called when a component
unmounted since `getTarget` was a dependency in the useEffect for resize
events.

The useResizeObserver was updated to now support mutable ref objects out
of the box so that it'll update the target after initial mount as well
as now moving the `target` into a getDerivedStateFromProps type of
lifecycle so that arrow functions can be used without causing the
useEffect to be cleaned + remade during multiple renders.