ResizeObserver Behavior Fixes and Changes
Updated the useResizeObserver to use `target` instead of `getTarget`.

Stopped using the "shared" ResizeObserver for additionl performance
since it lead to memory leaks when components unmounted. For some
reason, the `observer.unobserve` doesn't actually stop the handlers from
being called if it happened during a React unmount phase? The only way I
could prevent the memory leak warnings was to isolate the resize
observer to each useEffect and use `observer.disconnect()` which stops
*all** (1) observered targets. Using `observer.unobserve` would still
cause the memory leak warnings.

The downside to this is that it is less performant, but this was
probably a pre-mature optimiziation anyways.
https://github.com/WICG/ResizeObserver/issues/59