Huge build speed performance boost refactor
I haven't tested a full clean build yet but the next TravisCI build
should check it for me, but the build speed for an incremental build
changed from ~390s to ~170s.

I realized that I didn't need to type check everything en the commonjs
exports or the scss variables since it's already being checked in the
esmodules build. The type checking is super expensive especially when
duplicated. This commit updated all these files to enable the
`skipLibCheck` flag which is magic.

Also did a small tsconfig change so that the `target` is inherited from
the base `tsconfig.cjs.json` instead of re-defining it in each package's
tsconfig.cjs.json file.