Started Work for SassDoc
Switched up how the types folder behaves in the documentation size to
work a bit nicer and not needing the types prefix.

Removed some old and unused files.

Fixed the sassdoc.d.ts comments when I ran the vim's textwidth
formatter. Forgot that it doesn't understand code blocks and joined
those lines together.

Started the initial work for re-generating the SassDoc throughout the
site. This is really leading up to the stashed table rework I have so I
can test the new table API a bit more to see how much more convenient it
is.

Remaining work for the SassDoc:
- Complete the Mixin and Function compilations
- Write separate files for SassDoc by package and lazy load them as
  needed
- Copy or symlink the sassdoc.d.ts and sassdoc-custom.d.ts into the
  documentation package
- Finish the new SassDocPage component that is stashed along with the
  table rework

Next: Move on to expansion panels and check again what's left for v2
beta release.