Added better keyboard focus indication
Updated the states package so that there is also a default blue
box-shadow applied when an element is keyboard focused. Using only a
background-color change is not enough for a11y so adding the
outline-style back in a "prettier" way fixes this issue.

This behavior is opt-out though if this sort of thing isn't important
for your app.

Also fixed the layout package's `<main>` element's shadow to be updated
based on the side nav's appearance.