Fixed the link keyboard focus styles
I forgot that links are styled inline and the whole `::before` pseudo
element with box-shadow doesn't work well for those when they line wrap.
I could have fixed it by switching to `display: inline-block`, but that
looks pretty ugly most of the time when line-wrapping in the middle of a
link would have worked.

Switched to styling the `outline-width`, `outline-color`, and
`outline-style` for links instead.

Also added a sanity pre-commit hook that runs `yarn dev-utils variables`
whenever a `_variables.scss` file changes so that I can auto-commit
those. I included the changes for the layout package and the new
states/links variables now.