Added initial generated sassdoc and small sandbox fixes
Finished the setup for how sassdoc should work and started working on
creating the sassdoc pages (uncommitted).

Updated the documentation constants to be separate files and generate
the package names instead of reading it from the package.json. Reading
it from the package.json is kind of invalid when setting the "rootDir"
in typescript and there isn't a way to trim the unneeded json. This
broke some sandboxes that imported this constants file.

Updated the sandboxes after the constants changes and some missing
changes beforehand.

Updated the github urls to use the latest git commit sha instead of
master along with update the next.config.js to define the latest git
commit sha.