Fixed Firefox <select> Focus Styles
This is an annoying fix since there isn't really a nice way to style the
`:-moz-focusring` effect. You have to change the `color` value but that
also changes the color of the select's value... There's a workaround by
setting the `text-shadow` to a "guess" of the current color for this
flow. I don't know how many people will be using the NativeSelect
though.

Also updated the utils variables to make the focus-fix behavior
overridable at a global level instead of only at the
`rmd-utils-no-focus-outline` mixin level.