Implemented a quick way to override default icons
This became such a hassle that I decided to implement this for initial
v2 release. All icons throughout the app can now be updated by using
either the:

- IconProvider (@react-md/icon)
- Configuration's `icons` prop (@react-md/layout)

The whole pattern of having to create a new component that just changed
the specific `icon` prop to an SVG icon (or other) was super annoying so
I wanted a quick way to do it.

All components use the `useIcon` hook internally which will be a growing
list of default supported icons. It _might_ be nice to allow additional
icons in the future if it is somewhat useful for non-react-md
components, but probably overkill.

Also fixed an `id` issue for the theme toggle checkbox on the theme
builder page which caused the button in the header to gain the hover
styles when the checkbox was hovered (yay label + id stuff)