Added an UnstyledButton component
Added a simple UnstyledButton that renders a <button> element without
the default styles. This is really for the expansion panel stuff coming
up, but it is a nice component to have.

Updated the `rmd-states-focus-shadow` mixin to support creating the
`::before` or `::after` tag automatically as well since sometimes you
don't want all the states surface styles (background color changes).

Also updated the sassdoc to reflect some newer variables.