Fixed scss comment formatting
I wish comments were supported by prettier so this could be automated,
but it looks like it'll still be awhile for that. However, I finally
figured out how to update my vim config to support sassdoc triple slash
comments for the built-in formatter and used that.

Also updated the sheet's static z-index for some of the table work that
should hopefully be committed soon.