Updated Goals and Fixed Table for RTL
Updated the goals on the homepage of the documentation site to hopefully
be more concise and give an answer to what this library does without
needing to search through other pages.

Gave a detailed pros and cons list for using this library.

Fixed the table header styles so that it correctly right-aligns in RTL
languages. Also updated the custom align classes to be reflected in RTL.