Added Expansion Panel Package
Added the expansion panel package back for v2.

Mostly done, just need to write some tests for the usePanels hook and
some simple snapshots for the components.

Also need to figure out if the secondaryLabel functionality from v1 was
super important. That part was a pain since it needed to support
mobile which never worked well...