Fixed AutoComplete
Fixed the AutoComplete to no longer use the getEmptyValueData prop since
it caused infinite re-renders sometimes. Instead, created a boolean that
will allow calling the filter function even while the value is empty
instead so that a custom filter function can do this instead.

Also fixed the AutoComplete to set the listboxWidth back to "equal"