Chip Work
This is finishing up the work required for chips. This added some new
states, a selection state, an "activated" state, general transition for
a selected icon, and some demos.

Need to finish up the demos and clean up ones that aren't needed.

Also updated the IconProvider to use a selected icon.