Updated Custom Keyboard Behavior to stopPropagation
After playing with this some more, it seems better to stop propagation
after custom keyboard movement.

Also updated the AutoComplete to only call stopPropagation when the
escape key is pressed and there is a value or a listbox visible. This
will allow for the escape key to still close a dialog from an
autocomplete.