Added additional customization to Password component
Added a getVisibilityIcon prop to dynamically return an icon for the
visibility button based on the current type of the text field.

Added an onVisibilityClick handler to the button if that needs to be
hooked into.

That being said, it's probably simpler just to implement a custom
`Password` component instead by this point since the built-in `Password`
component is only like 5 lines of additional functionality excluding
this customization.