Fixed GridList on higher pixel density screens
Not sure if that is _really_ what the cause was, but the columns really
only seemed to be off on 1440p and 4k monitors as well as custom zoom
levels.

`offsetWidth` (and `offsetHeight`) are actually rounded values of the
element's width (`Math.ceil`). This causes problems if the element's
width is barely greater than the previous number since it'll get rounded
up. IE: `712.023` would get rounded to 713 in `offsetWidth`.

This normally wouldn't be much of a problem, but now the width is off by
almost a full pixel which means that the cells won't be able to fill the
container's width and basically show an empty cell on each row.

Switching to `element.getBoundingClientRect().width` fixes this issue
since it has better width precision and does not round.