Fixed useOrientation for browser support
Welp... I should have realized that window.screen.orientation might not
be fully supported when I had to update a test to polyfill it and not
crashing. Anyways, most mobile browsers don't support it and neither
does Safari so updated useOrientation to have a really simple polyfill
that falls back to checking the available height.