Fixed HighlightMatches demo for iOS
Setting `height: 100%` for the phone apparently does not work in some
browsers if the parent element does not also have height set. Since the
content is for this example wrapped in a full-height container from
flexing, the height was practically 0 since it was not specified.

Since Safari is a great browser and setting `overflow: auto` also means
fixed elements will be hidden, it was impossible to exit the demo in
iOS.