Updated positioning logic again
The fixed positioning hook will now attach the scroll and resize
listeners immediately after the element becomes visible instead of
relying on another render (it forces a re-render when node changes now)

Updated the positioning logic so that when using position absolute, the
current page scroll is subtracted when not in a dialog. Probably will
need more fixes for that later.

Also updated HighlightMatches after the autocomplete changes.