Updated TOC and implemented Site Search
Created a new dev util to "index" the documentation site which creates
static table of contents for all pages that should have quick links. In
addition, it creates a searchable list of metadata about each route
within the documentation to populate the search results.

I also implemented the first version of search within the documentation
site with the AutoComplete. I think the AutoComplete will need to be
reworked a bit in the future since right now it is annoying to create a
custom rendered item and I can't do nice things like adding items that
aren't searchable, but are used for display purposes like loading or
waypoints to fetch more results.

I also realized I need to add the additional functionality to:

- control the value manually
- have hooks for when the search results are visible
- other stuff...