chore: last attempt build speed improvement
This will be the last attempt for awhile using typescript as a compiler
since I've wasted enough time playing around with this. It'll probably
be better to switch to babel for compilation especially since I'll be
able to use additional plugins (like remove prop-types automatically)

Switched up the build to use project references and the new --build
flag. I also created a new `dev-utils configs` script to update all the
tsconfigs in each package to handle this. There isn't too much a
difference running it locally, ~3-4 min for a clean build instead of the
5-6 min, but the travis build machine seems to have gone from 20 minutes
to 8-10 minutes (including tests).. so that's kind of nice. However, the
local build and watch behavior is a lot more performant and quicker. A
rebuild is about 10 seconds instead of 2 min and spinning up only 2 tsc
watchers for all of react-md instead of 24 (x2) makes computer not die.

Also updated the `prepublish`/`build` behavior so that `yarn build` will
now also copy the styles to the dist directories. The `prepublish`
script from `dev-utils` is really just to get some additional logging
about the library size for the UMD bundle and min/max css bundle size.