fix(menu): Escape and Tab Keyboard Behavior
Pressing the Escape key will now correctly stop propagation for menus.
This is so that having a menu within a dialog will not also close the
dialog immediately.

Fixed the tab behavior for portalled menus so that the default tab
behavior is disabled and the menu button is focused instead. Portalling
ruins tab index behavior so either the first focusable element in the
page would be focused on a tab keypress or the last focusable element.
Focusing the menu button "fixes" this behavior, but would still require
another tab keypress to focus the next or previous element.