feat(app-bar): updated app-bar height functionality
Updated the `AppBar`'s height to only be a single prop: `height`. It
started becoming confusing with the three different props for sizing
(`dense`, `prominent`, `derived`) and how they actually work together.
The new height system is just:

- none
- normal (default)
- dense
- prominent
- prominent-dense

This also added a new prop `flexWrap` which defaults to `true` when the
height is set to `none`, `prominent` or `prominent-dense` to allow
flex-wrap within the AppBar