fix(focus): Fixes for keyboard focus behavior
The `Dialog` component now has a prop to conditionally disable the tab
focus wrap behavior since it helps with the sheet.

The sheet now correctly sets itself to `hidden` when it does not unmount
while invisible so that its contents cannot be tab focusable. While
writing this out I realized I might need to update the
getFocusableElements to exclude items that are a child of `hidden`...

Updated the `rmd-utils-base` mixin to add a fix for hidden elements to
always be display none and removed that fix from the collapse styles.

Fixed the keyboard focus behavior in the documentation site's
DemoSandbox so that it doesn't get stuck in the navigation tree. Also
made it so that the code can now be scrolled correctly with the
keyboard.