feat(menus): Changed default closing behavior
After playing around with this a lot more, I started to disagree with
the default behavior of hiding if the user scrolls the page or the
browser is resized. The fixed positioning should behave as closely to a
`position: relative` + `position: absolute` as possible by default and
the developer can opt-in to this new functionality if their UX team
decides that is what's preferred.

This changed all the `disableCloseOnResize` to `closeOnResize`,
`disableCloseOnScroll` to `closeOnScroll` and updated the autocomplete
to have the same naming conventions: `disableHideOnResize` ->
`closeOnResize` and `disableHideOnScroll` -> `closeOnScroll`