chore(alert): @react-md/alert changelog and fixes
Added the changelog for the alert package

Fixed some of the typedoc for the message queue context.

Added a small padding fix for the stacked messages with an action
button.

Added in some default dark/light theme colors for the alert even though
they are the same right now. This will need to be fixed once I figure
out how to do the "raising dark theme elevation" feature since there
isn't enough contrast right now between the background and the toast in
dark themes.