fix(transition): Fixed React Fast Refresh Issue
Next.js v9.4 has started using the new
[Fast Refresh](https://nextjs.org/blog/next-9-4#fast-refresh)
functionality in React. Unfortunately, this seems to completely break
transitions use it relies on skipping the transition on the first render
and triggering a `useEffect`. The fast refresh does not reset the
`useRef` for isFirstRender and the `useEffect` is triggered again which
causes the transition to fire again.

This actually caused the app to crash since a "requirement" for a css
transition is to have access to a DOM node to force a DOM repaint, but
those are actually set to `null` for a fast refresh and causes the app
to crash. I could probably re-write the whole `useTransition` and
`useCSSTransition` to get around this, but the simplest solution is to
no longer throw an error if there is not a DOM node and "skip" the
transition callbacks for this flow.

I tried a couple of things like not throwing an error if the
`window.$RefreshHelpers$` exists and instead returning the
`document.body`, but that's even hackier and `window.$RefreshHelpers$`
looks like it's restricted to nextjs only and other implementations of
fast refresh might cause this to break as well.