fix(dialog): Fixed full page dialog styles
Updated the dialog to always set max-height and max-width instead of
only when it is centered. It makes it much easier to switch between
full-page and centered dialogs without any fixed headers which is a
common pattern when there is a lot of content within a dialog and
switching between mobile and larger devices.