chore: Updates to look better in npm
Created a new `yarn dev-utils rmd-readme` script that will copy the root
README.md to the `react-md` folder and insert some additional
information into the README. This is just to help add documentation if
this package is found through npm instead of GitHub or other sources.

Also updated all the package.json files to link to their specific
documentation page as the `"homepage"` field.