chore(docs): Removed API Routes
I haven't figured out how to generate the API docs yet and it's better
to remove these pages so it isn't included in search or lead to
confusion since there's nothing there.

Hopefully I'll be able to figure out typedoc, react-docgen-typescript,
or react-docgen (with the new typescript support) soon so it can be
added back.