fix: gzip size for package
Not sure how I managed to mess this up so badly.. but the gzip size
package does not parse by file path but instead the string that was
provided... So the current v2 UMD size is way too big with including all
the icons by default. Going to split up the UMD bundles next.