fix: Scroll active element into view while focusing
It looks like using element.focus() doesn't always scroll the element
into view if a parent element has scrolling enabled. This caused a
problem with large dropdown menus and using the ArrowUp key to trigger
the menu that would not have the current element visible until the
ArrowUp key was pressed again.