fix(theme): Fixed `rmd-theme-get-swatch` to loop over all `rmd-theme-colors` instead of the primaries only
I probably should have added some other `@if`s to this after the
looping to ensure a color has been found, but the checks before this
point should really catch the other cases.

Because this only looped over the primary colors, if you set the
`$rmd-theme-primary: $rmd-green-500;`, this color would be resolved as
`$rmd-green-500-300` in the chip package for the
`$rmd-chip-themed-background-color` variable which threw an error when
trying to do color contrast ratio fixes in `$rmd-chip-themed-color`.

Closes #884